Update to the latest tatami interface. (#6)
We now handle differences in the TileDB attribute/dimension type compared to
the tatami interface types. This requires a lot of work to test that the
dimensions are computed correctly for a variety of signed/unsigned index types.

We improved the oracular caching to consolidate slabs into a single request to
the TileDB library. We switch to a variable slab cache for sparse oracular
extraction to improve cache utilitization. We also replace the SoloCore class
as we can achieve the same memory usage by reusing the other Cores with a chunk
length of 1 and a maximum number of cached slabs to 1.

We introduce a dedicated serialize() function to document and consolidate the
locks around the TileDB library. We actually need to implement a lock during
the parallelization tests because it occasionally deadlocks otherwise.

Updated the CI test workflow to use the latest versions, avoid using ctest,
and to pass along the Codecov token to report coverage.

Finally, updated the README.